Removed float in forward and added it to mem_out

+++ network/base_network.py
@@ -22,7 +22,7 @@ class Plastic(ABC, nn.Module):
         self.total_drifts = total_drifts
 
     def forward(self, x, global_step=None):
-        return self._forward(x.float() / 255.0, global_step)
+        return self._forward(x / 255.0, global_step)
     
     @abstractmethod
     def _forward(self, x, global_step):
@@ -139,7 +139,7 @@ class Large_SNN(Plastic):
         mem_fc = self.lif_fc.init_leaky()
         mem_head = self.lif_head.init_leaky()
 
-        mem_out = torch.zeros(x.size(0), self.action_space).to(x.device)
+        mem_out = torch.zeros(x.size(0), self.action_space).to(x.device).float()
 
         spk1_average = 0
         spk2_average = 0
